Move all netcdf writer tests into one file
